feat: implementada request/response com XML para o TransactionType

+++ app/Services/TransactionTypeService.php
@@ -23,16 +23,31 @@ class TransactionTypeService
     /**
      * Envia para o TransactionTypeRepository os dados para criar uma nova instância de TransactionType
      * @param array $data
-     * @return \App\Models\TransactionType
+     * @return array|mixed
      */
-    public function store(array $data)
+    public function store($data)
     {
-        return $this->repoTransactionType->store($data);
+        try {
+            // TODO: Tratar os dados da requisição, antes de chamar o repoTransactionType->store
+            $transactionType = $this->repoTransactionType->store([
+                'type' => $data['type'],
+            ]);
+
+            $status = 200;
+
+            return compact('transactionType', 'status');
+
+        } catch (\Throwable) {
+            $message = 'Transaction Type not stored';
+            $status = 400;
+
+            return compact('message', 'status');
+        }
     }
 
     /**
      * Retorna todas as instâncias de TransactionType do banco de dados
-     * @return array[\App\Models\TransactionType]
+     * @return \Illuminate\Database\Eloquent\Collection<int, static>
      */
     public function getList()
     {
@@ -42,31 +57,83 @@ class TransactionTypeService
     /**
      * Retorna uma instância de TransactionType a partir do id informado
      * @param mixed $id
-     * @return \App\Models\TransactionType
+     * @return array
      */
     public function get($id)
     {
-        return $this->repoTransactionType->get($id);
+        try {
+            $transactionType = $this->repoTransactionType->get($id);
+            $status = 200;
+
+            throw_if($transactionType == null);
+
+            return compact('transactionType', 'status');
+
+        } catch (\Throwable) {
+            $message = 'Transaction Type not found';
+            $status = 404;
+
+            return compact('message', 'status');
+        }
     }
 
     /**
      * Atualiza os dados de uma instância de TransactionType
      * @param array $data
-     * @param mixed $id
-     * @return \App\Models\TransactionType
+     * @param int|string $id
+     * @return array|mixed
      */
-    public function update(array $data, $id)
+    public function update($data, $id)
     {
-        return $this->repoTransactionType->update($data, $id);
-    }
+        try {
+            // TODO: Tratar os dados da requisição, antes de chamar o repoTransactionType->store
+            $keys = [];
+            $values = [];
+            foreach ($data as $key => $value) {
+                array_push($keys, $key);
+                array_push($values, $value);
+            }
 
+            $processed_data = array_combine($keys, $values);
+
+            $status = 200;
+
+            $this->repoTransactionType->update($processed_data, $id);
+            $transactionType = $this->repoTransactionType->get($id);
+            $status = 200;
+
+            throw_if($transactionType == null);
+
+            return compact('transactionType', 'status');
+
+        } catch (\Throwable) {
+            $message = 'Transaction Type not found';
+            $status = 404;
+
+            return compact('message', 'status');
+        }
+    }
     /**
      * Remove uma instância de TransactionType do banco de dados
-     * @param mixed $id
-     * @return mixed
+     * @param int|string $id
+     * @return array
      */
     public function destroy($id)
     {
-        return $this->repoTransactionType->destroy($id);
+        $transactionType = $this->repoTransactionType->get($id);
+
+        if ($transactionType == null) {
+            $message = 'Address not found';
+            $status = 404;
+        } else if (count($transactionType->transactions) > 0) {
+            $message = 'This Transaction Type has associated transactions';
+            $status = 405;
+        } else {
+            $this->repoTransactionType->destroy($id);
+            $message = 'Transaction Type destroyed';
+            $status = 204;
+        }
+
+        return compact('message', 'status');
     }
 }
